requirements.txt
face_recognition==1.3.0
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
numpy==1.26.0
pillow==10.0.1
sqlalchemy==2.0.23
pydantic==2.5.2
